+++ DataMapping/Data/DT_User_Role.cs
@@ -20,7 +20,7 @@ namespace DataMapping.Data
             DataColumn[] PK_COLUMN = new DataColumn[2];
             this.TableName = TABLE_NAME;
             PK_COLUMN[0] = this.Columns.Add(USER, typeof(string));
-            PK_COLUMN[1] = this.Columns.Add(IDROL_COLUMN, typeof(int));
+            PK_COLUMN[1] = this.Columns.Add(IDROL_COLUMN, typeof(decimal));
             Columns.Add(ROLE, typeof(string));
             this.PrimaryKey = PK_COLUMN;
         }
